Workflow is in beta for 5.x!
Well, crack open the champagne and put your party hats on. After hours of hard work by drewish, webchick, samo, moshe, bloomaniac, jvandyk, Egon Bianchet, unleached, mr. dale, bdragon, and myself, workflow is running on Drupal 5. Special thanks to all those listed above who took the time to patch, review, critique, and otherwise move the ball towards the endzone.

Now, we're down to the final yards (can you tell the Super Bowl is tomorrow?). This is a beta release (perhaps the first of several). This means the code in unstable, buggy (read down for some known bugs), and will be changing in the immediate future. That said, this is mostly a compatibility release, so not much has changed in the code base. I am confident the full release will be out soon.

I'm putting out a call for people interested in testing this release. Please put it through it's paces. I'm especially interested in testing of:
- integration with actions (make sure APIs haven't changed in subtle ways)
- views integration (test your views, make sure they still work)
- upgrades (*** do not test on your production systems, make copies! but then see if the upgrade is successful)
- translations (these have fallen out of date. translators are very welcome to submit changes/updates)

If you want to jump right in, here's my (ever growing) list of bugs:
- when deleting a workflow, <em> is escaped into the drupal_set_message
- on workflow form, remove the label from the select list. hight current state better
- on workflow form, authorship still not checked right!
- scheduling puts extra message into logs
- scheduling is ignored and done immediately on workflow tab. :-(